// src/types/inventoryVoucher/GstBillTypes.d.ts
export interface GstBillFormData {
  date: Date;
  billNo: string;
  partyId: string;
  address: string;
  mobileNo: string;
  gstin: string;
  gstRate: string;
  itemName: string;
  itemUnit: string;
  itemQuantity: string;
  itemRate: string;
  itemHsn: string;
  discount: string;
}
// ...
interface ItemData {
  item_name: string;
  item_qnty: string;
  item_unit: string;
  item_hsn: string;
  item_rate: string;
  item_tot: string;
}
// ...
export interface InvoiceData {
  Sales_Id: 6;
  Sales_Date: string;
  Sales_No: string;
  Party_Name: string;
  Gross_Amt: number;
  Cgst_Rate: number;
  Cgst_Amt: number;
  Sgst_Rate: number;
  Sgst_Amt: number;
  Igst_Amt: number;
  Round_Amt: number;
  Discount: number;
  ItemData: {
    Item_Name: string;
    Item_Qnty: number;
    Item_Unit: string;
    Item_Rate: number;
    Item_Hsn: number;
    Item_Tot: number;
  }[];
}

export interface GstBillTableData {
  Id: number;
  Sales_Id?: number;
  Sales_Date?: string;
  Sales_No?: string;
  Bill_Date?: string;
  Bill_No?: string;
  Party_Name?: string;
  Gross_Amt?: number | string;
  Total?: number | string;
  Cgst_Amt?: number | string;
  Sgst_Amt?: number | string;
  Igst_Amt?: number | string;
  Round_Amt?: number | string;
  Discount?: number | string;
}

export interface ItemTableData {
  itemName: string;
  itemUnit: string;
  itemQuantity: string;
  itemRate: string;
  itemHsn: string;
  itemTotal: string;
}